# brainX_back/Intelligence-Service/src/main/resources/contracts/knowledge-intelligence.openapi.yaml
openapi: 3.0.3
info:
  title: BrainX API SSOT - Knowledge Intelligence
  version: 1.0.0-revised
  description: 'BrainX public and internal REST contract.


    Revision policy:

    - Public/client API remains under /api/v1.

    - Service-to-service synchronous APIs are explicitly isolated under /internal/v1.

    - Duplicate member-deletion cancel endpoint was removed.

    - /support/inquiries was consolidated into /support/tickets.

    - Token usage recording is event-first via AsyncAPI; no public REST command is exposed.

    - Publish jobs return 202 Accepted when status is PENDING.

    - Operation-level consumer metadata is expressed through x-consumers.

    - Operation-level event coupling is expressed through x-produces-events and references the AsyncAPI SSOT.'
servers:
- url: http://localhost:8080
  description: Local API Gateway
- url: https://api.brainx.com
  description: Production API Gateway
...
paths:
...
  /api/v1/ai/models:
    get:
      operationId: listAiModels
      summary: 사용 가능한 AI 모델 목록
      tags:
      - Knowledge Intelligence
      x-producer-service: knowledge-intelligence
      x-visibility: public
      x-consumers:
      - id: web.chat
        type: frontend
        description: RAG chatbot UI
      - id: web.settings
        type: frontend
        description: Account, profile, security and AI settings
      x-produces-events: []
      x-internal-sync-calls: []
      x-async-boundary: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiSuccessBase'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      $ref: '#/components/schemas/AiModelsData'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearerAuth: []
  /api/v1/ai/model-settings:
    put:
      operationId: putAiModelSettings
      summary: AI 모델 설정 변경
      tags:
      - Knowledge Intelligence
      x-producer-service: knowledge-intelligence
      x-visibility: public
      x-consumers:
      - id: web.settings
        type: frontend
        description: Account, profile, security and AI settings
      x-produces-events:
      - AiModelSettingsChanged
      x-internal-sync-calls: []
      x-async-boundary: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiModelSettingsPutRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiSuccessBase'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      $ref: '#/components/schemas/AiModelSettingsData'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearerAuth: []
...
  /api/v1/users/me/style-profile:
    get:
      operationId: getStyleProfile
      summary: 문체 프로필 조회
      tags:
      - Knowledge Intelligence
      x-producer-service: knowledge-intelligence
      x-visibility: public
      x-consumers:
      - id: web.settings
        type: frontend
        description: Account, profile, security and AI settings
      x-produces-events: []
      x-internal-sync-calls: []
      x-async-boundary: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiSuccessBase'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      $ref: '#/components/schemas/StyleProfileData'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearerAuth: []
    put:
      operationId: putStyleProfile
      summary: 문체 프로필 설정
      tags:
      - Knowledge Intelligence
      x-producer-service: knowledge-intelligence
      x-visibility: public
      x-consumers:
      - id: web.settings
        type: frontend
        description: Account, profile, security and AI settings
      x-produces-events:
      - UserStyleProfileChanged
      x-internal-sync-calls: []
      x-async-boundary: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StyleProfilePutRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiSuccessBase'
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      $ref: '#/components/schemas/StyleProfileData'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ApiSuccessBase:
      type: object
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
      required:
      - success
      - message
      additionalProperties: false
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum:
          - false
        data:
          nullable: true
        message:
          type: string
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            traceId:
              type: string
            details:
              type: object
              additionalProperties: true
              description: Machine-readable error details
          required:
          - code
          - message
          additionalProperties: true
      required:
      - success
      - message
      additionalProperties: false
...
    AiModelsData:
      type: object
      properties:
        models:
          type: array
          items:
            type: object
            properties:
              modelId:
                type: string
              name:
                type: string
              provider:
                type: string
              vendorInputCostPer1kTokens:
                type: number
                nullable: true
              vendorOutputCostPer1kTokens:
                type: number
                nullable: true
            required:
            - modelId
            - name
            - provider
            additionalProperties: false
        enabledModels:
          type: array
          items:
            type: string
        costInfo:
          type: object
          additionalProperties: true
      required:
      - models
      - enabledModels
      additionalProperties: false
    AiModelSettingsPutRequest:
      type: object
      properties:
        defaultModelId:
          type: string
        userApiKeys:
          type: object
          additionalProperties: true
      required:
      - defaultModelId
      additionalProperties: false
    AiModelSettingsData:
      type: object
      properties:
        settings:
          type: object
          additionalProperties: true
      required:
      - settings
      additionalProperties: false
...
    StyleProfileData:
      type: object
      properties:
        conversationTone:
          type: object
          additionalProperties: true
        writingStyle:
          type: object
          additionalProperties: true
        assistanceStyle:
          type: object
          additionalProperties: true
        detectedFromNotesAt:
          type: string
          format: date-time
          nullable: true
      required:
      - conversationTone
      - writingStyle
      - assistanceStyle
      additionalProperties: false
    StyleProfilePutRequest:
      type: object
      properties:
        conversationTone:
          type: object
          additionalProperties: true
        writingStyle:
          type: object
          additionalProperties: true
        assistanceStyle:
          type: object
          additionalProperties: true
      additionalProperties: false
...